Don't return original PSBT if input contribution fails
If input selection or contribution fails, either broadcast the original PSBT or return
`unavailable`. Since payjoin-cli is a reference implementation of a non-automated receiver,
I prefer to return `unavailable` since the sender can choose to broadcast that original PSBT.
It makes less sense to me to return the original PSBT without modifications which will likely
cause the sender to broadcast the original despite no payjoin happening. Let the sender make
that choice.