Add additional tests for send v2 module
This is an attempt to both add some additional test vectors and to create some
headway on the tests deficiencies caught by cargo mutants.

Currently missing is the test for process_response which I was unable to make
significant progress on as that test currently expects an actual response from
an ohttp relay. Until we can make any progress on that test I have left that
test absent for now.